requirements.txt
Click==7.0
flask==1.0.2
flask-jwt-extended==3.14.0
cryptography==2.3.1
itsdangerous==0.24
flask-caching==1.4.0
cherrypy==18.0.1
wsgi-request-logger==0.4.6
Flask-SQLAlchemy==2.3.2
Flask-Dance==1.1.0
Flask-Dance[sqla]==1.1.0
Flask-Login==0.4.1
simplejson==3.15.0
sqlalchemy==1.2.15
sqlalchemy-utils==0.33.9
sqlalchemy-json==0.2.1
python-dotenv==0.9.1
grpcio==1.17.1
apispec==0.39.0
marshmallow-sqlalchemy==0.15.0
chaosplatform-grpc==0.1.0
chaosplatform-relational-storage==0.1.1
chaosplatform-scheduling==0.1.0
chaosplatform-scheduler==0.1.0
chaosplatform-auth==0.1.0
chaosplatform-account==0.1.0
chaosplatform-experiment==0.1.0
toml==0.10.0